enforced integer division for indexing
Python 3 introduced "true" division and can produce float values from two integers unless directly type cast or if the floor division operator is used.  I propose using the latter when computing the half_length variable as it is required to be of type int when used to slice quality in C33 check.